update take_screenshot to work with urls as well as rfiles

+++ R/makeExamples.R
@@ -46,13 +46,18 @@ create_chart <- function(rFile){
 #'
 #' @param path to R file containing code to create an rChart
 #' @param imgname name of the plot to save to
-take_screenshot <- function(rFile, imgname = 'plot1', delay = 10000, upload = F){
-  rCode = paste(readLines(rFile, warn = F), collapse = "\n")
+take_screenshot <- function(src, imgname = 'plot1', delay = 10000, upload = F){
+  if (file.exists(src)){
+    rCode = paste(readLines(src, warn = F), collapse = "\n")
+    chart = source(src, local = TRUE)$value
+    chart$set(width = 600, height = 325)
+    tf <- tempfile(fileext = ".html"); on.exit(unlink(tf))
+    chart$save(tf)
+  } else {
+    tf <- src
+  }
+  
   script = system.file('utils', 'screenshot.js', package = 'rCharts')
-  chart = source(rFile, local = TRUE)$value
-  chart$set(width = 600, height = 325)
-  tf <- tempfile(fileext = ".html"); on.exit(unlink(tf))
-  chart$save(tf)
   cmd1 <- sprintf('casperjs %s %s %s %s', script, tf, imgname, delay)
   system(cmd1)
   # system(sprintf("convert %s.png  -resize 288x172", imgname))
